Doc: Use local value for \QMLD macro in Qt Design Studio Manual
This enables us to get rid of the product name "Qt Quick Designer"
in Qt Design Studio context. However, the macro can no longer be
used for UI text. Actually, it does not make much sense, because the
UI text is not dynamic. It would be good to also replace it in the
UI with something more generic, or with "Form Editor", for example.

Change-Id: I5884f090f915d5ff0c64428239bfd324cd1941c2
Reviewed-by: Brook Cronin <brook.cronin@qt.io>
Reviewed-by: Thomas Hartmann <thomas.hartmann@qt.io>