QbsProjectManager: Fix install step
We got the install command name wrong, which caused the installation to
fail. In addition, we forgot to add handling for the "protocol-error"
message from qbs, so the step was hanging, rather than aborting.
Finally, we triggered a number of assertions in
BuildStep::buildConfiguration().
This went all unnoticed for a while, because the install step is not
enabled by default these days.

Change-Id: I906e7e472563d4ad8fc7557bd706a7cb67f9f2ba
Reviewed-by: hjk <hjk@qt.io>