CppEditor: Fix crash with adjacent raw string literals
Our lexer is not good at handling newlines embedded in raw string
literals; basically, it sees every continuation line as a new string. So
if such a continuation is followed directly by a new raw string literal,
we have to take care to tell them apart properly. In particular, it can
happen that an end delimiter occurs before an opening delimiter.

Fixes: QTCREATORBUG-24577
Change-Id: I631d0617d85e91f49a25d309d53778da0170eb3b
Reviewed-by: David Schulz <david.schulz@qt.io>