AutoTest: Fix marking for removal and sweeping of Quick Tests
If opening and editing a QML file we did no more mark tests of this
file as to be removed. In turn we may have got "new" items when e.g.
just modifying the name of a function or test case and but did not lose
the old one.
Broke with 90fd6870c5a.

Change-Id: I86917089ffb9d4acb2f735f2f3e0f112d8cdc65a
Reviewed-by: David Schulz <david.schulz@qt.io>