QmlDesigner: Refactor preview image data caching
Made image cache always be local to the NodeInstanceView and changed
the key to always be the node id to avoid wrong data to be shown
on tooltip in some cases, such as id change or same image used by
multiple nodes.

Also clean up old cached item on id change.

Change-Id: I6fedea2a158d62e03f9d04be4cd26ae34550d09c
Fixes: QDS-2843
Reviewed-by: Thomas Hartmann <thomas.hartmann@qt.io>