Project Tree: Delay syncing with editor when using the context menu
When "Synchronize with Editor" is enabled, every focus change in the
application causes the current node in the project tree to be reset to
the one corresponding to the current document. In order to still be able
to use the context menu, there is a flag in the project tree suppressing
this behavior for as long as the context menu is open. However, some
actions offered by the context menu open an additional dialog, and it is
confusing to users if the current node changes at this point.
We therefore extend abovementioned hack so that the suppression of
automatic syncing is extended for the lifetime of the function called
from the context menu.

Fixes: QTCREATORBUG-24699
Change-Id: I209150aee76e534a966efc4d0afe5261d6dcd521
Reviewed-by: hjk <hjk@qt.io>