QMakeProjectManager: Do not ask to rebuild too often
Due to the incomplete aspectification here, some of the connections
are set up in createConfigWidget(). Config widgets are created
and destroyed when the build configuration changes, and connections
that did not directly depend on the widget or subwidgets accumulated.

Use the the created widget as guard object to ensure connections
live only as long as the widget.

Task-number: QTCREATORBUG-24725
Change-Id: I4581f98e684ac92fd43ceba598d391c4dc8dbebe
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>