AutoTest: Prefer active target over further
If there are more than one target for a run configuration
we ended up in the first matching.
Tests may use different environments to test different things.
If the user tried to handle this by using a modified
environment for different targets failed even when the
correct target was selected.

Change-Id: Ie954c985969ffd187fe60b4ecb86aa2717d559f1
Reviewed-by: David Schulz <david.schulz@qt.io>