EditorManager: Avoid changing the editor in between when closing editors
Do not call setCurrentEditor(nullptr) since this sends a
currentEditorChanged signal which claims that no editor is active, which
can lead to unnecessary updates.
In the specific case, the git branch view was updating with the project
directory as a new repository, and then, after the actual new current
editor was set updated again with the repository of the editor.
This can be needlessly expensive if the previous and new editor are in
the same repository, but the project repository is something else, like
when having the Qt Creator super-repository open, or when editing files
unrelated to the project.

Change-Id: I515c3c40a86bfa2713ff13460ddfda974bc08c59
Reviewed-by: David Schulz <david.schulz@qt.io>