Revert "ClangFormat: Relax the clang format file a little bit"
That change does an experiment on the clang format file without
discussing the results before. It wasn't tested and discussed on a
larger code base, and ignores tests/manual/clang-format-for-qtc.

Individual ("pre-discussion") experiments with the clang-format file
should be done in a separate experimental version instead (e.g.
somewhere in dist/).

This reverts commit 009c2745e27a9b1e9bc3886f159c271fb491c7d4.

Change-Id: I5f9fa50d2ccc63807218ae5d193a3a0718956576
Reviewed-by: hjk <hjk@qt.io>