clangbackend: Work around wrong cursor info from libclang
Alias type declarations of in-place struct definitions are not reliably
getting their cursor exposed, so the surrounding context is reported
instead, which causes the wrong lexical parent to be deduced. To the
user, this manifests itself as a wrong hierarchy being displayed in the
outline.
We try to detect this condition and correct it manually.

Task-number: QTCREATORBUG-24875
Change-Id: Ib4e177efa9148697eddde986a91fdd5d7b272d76
Reviewed-by: Christian Stenger <christian.stenger@qt.io>