Compile fix in serialize_container when building with Qt 6
After bfc6e2d69d2514f7b4e889d08475972eb2f340a6 in Qt 6
we have a different signature of QList::append when QList
is instantiated with a pointer to type (T *). After the change
we get QList<T *>::append(T *), and before we got
QList<T *>::append(T * const &). This breaks the code in
serializer.

In addition, get rid of warning: "use of function template name
with no prior declaration in function call with explicit template
arguments is a C++2a extension". In order to fix it, we explicitly
declare the full function signature and let compiler choose the
right overload of attr() function template.

Change-Id: I4132a1307faa9a6bd1b585251a6bb5189a0b1764
Reviewed-by: Eike Ziller <eike.ziller@qt.io>