ProjectExplorer: Add shortcut to "Configure Project"
Note that this does not work with 'Return' immediately, as the focus
is usually on the left tree view in that situation. <Tab>, <Return>
work.

We can try to switch the focus by default in a separate patch if
that is wanted.

Task-number: QTCREATORBUG-20720
Change-Id: Iefbf7f2c1493947b847fd9c89e86b4de56a65543
(cherry picked from commit 08f3696f21d2ce69623ef01b5765b23d74bebb35)
Reviewed-by: Eike Ziller <eike.ziller@qt.io>