Designer: Fix escaping '&' again
It seems impossible to show a single '&' character
on UI elements, as this character is also used for
the mnenomics.

So having a single '&' followed by a space creates
a space character, having two of them ("&&") creates
an underlined '&' (seen on Windows and Linux).

So let's just work around this long-living issue
and avoid the abbreviation.

Amends commit bc37a9ab7d8

Change-Id: Ibcbfcac62d7be8c091ab2fc51621bddc9dad7e00
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>