QmlDesigner: Fix navigator preview tooltip issues
Preview tooltip is now shown as separate frameless tool window instead
of just a child widget to ensure it is always on top.

Fixed the layout issues with preview tooltip when the navigator
is moved to another screen with different pixel ratio.

Node internal id is used to identify the hovered node instead of
row index, as row index is only relative to immediate parent node.

Change-Id: Ida06565c7addfb6247e13ded11a1fed51097f0e9
Fixes: QDS-3119
Reviewed-by: Thomas Hartmann <thomas.hartmann@qt.io>