ProjectExplorer: Un-tighten the RunConfiguration::isConfigured() check
Only report a run configuration as unconfigured if it has a fundamental
problem that should inhibit automatic creation.
Adapt the potentially reported issue in QdbRunConfiguration accordingly.
The concrete reason is that for cmake projects, deployment information
is only available after deployment has happened, so the on-device
exutable is not known right away.

Task-number: QTCREATORBUG-24685
Change-Id: Ibe49649ebb001d3ba2b752c9c335687da9c08ce7
Reviewed-by: hjk <hjk@qt.io>