Don't call virtual method from c'tor
This patch fixes the following issues:

1. The c'tor of NonInternalLibraryDetailsController
called slotLibraryTypeChanged(), which called in turn
virtual updateWindowsOptionsEnablement(). This method
is reimplemented in subclasses, so we can't easily
make this method final.

The solution is to reorganize the code so that we don't
call anymore updateWindowsOptionsEnablement() from the
NonInternalLibraryDetailsController c'tor, but instead
we rely on the fact that it will be called indirectly through
updateGui, called for each leaf subclass from its c'tor.

2. Don't call multiple times updateGui when constructing
the subclass of NonInternalLibraryDetailsController or
when changing library type (from combo box). Before, each
time that happened, the updateGui was called three times.

3. Provide an ugly hack to nicely resize the wizard when
showing more widgets (if needed). This also eliminates flickering
between transitioning from Type to Details page.
It makes the layout much better, as it respects the reasonable
minimum size (when resizing to smaller size by the user).
The request to unhack addressed here: QTBUG-88666.

4. Mark all virtual methods with final for all
subclasses of LibraryDetailsController, where appropriate.

Change-Id: Ief7e5e9e1efdbfca5a06eaeeb1528018c29c39ca
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>