Fix qtcreator_process_stub on Windows
When building with CMake, this is compiled with `-DWIN32_LEAN_AND_MEAN`.
This reduces the number of headers that are included by default,
and lead to e.g. malloc and exit not being available:

..\src\libs\utils\process_stub_win.c(149): warning C4013:
'malloc' undefined; assuming extern returning int

And therefore reading the environment file was failing.
Include the needed stdlib.h.

Fixes: QTCREATORBUG-24829
Change-Id: I24722dfd17ecf0e4e6c9059e2cececbb5923a976
Reviewed-by: Orgad Shaneh <orgads@gmail.com>
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>