CMake: Improve handling of QML Debugging and Profiling option
The additional parameters to pass to CMake are set in a
"extraCMakeArguments" property, but we had two places where this was
set, overwriting each other: from the list widget where the user makes
changes, and from the QML debugging option.

Combine the two code paths. This has the additional advantage that the
changing QML Debugging and Profiling option behaves the same as changing
variables in the list - the "Apply Configuration Changes" button gets
enabled when changing the option, and one can use it to apply the
change.

Task-number: QTCREATORBUG-24988
Change-Id: I785c6804c9597a9eba471f56babaae966ce84b17
Reviewed-by: Cristian Adam <cristian.adam@qt.io>