Debugger: Fix editing breakpoints via breakpoints view
As the comment for the 'm_parameters' member
in the 'BreakpointItem' class says, that member is
supposed to hold the "Parameters acknowledged by engine",
which the parameters newly set in the dialog are not
yet at this point in time.

Therefore, if a global breakpoint exists, apply the
new parameters there, so they are considered as pending
changes (also returned by 'Breakpoint::requestedParameters')
when 'BreakHandler::requestBreakpointUpdate' is
called a few lines below, and new values get applied
properly.

Task-number: QTCREATORBUG-25035
Change-Id: I8bfd9123785ee911ff180d8311fa2559b55be580
Reviewed-by: hjk <hjk@qt.io>